# design/api/openapi.yaml
openapi: 3.0.3
info:
  title: IDaaS系统API接口文档
  description: |
    IDaaS（Identity as a Service）系统API接口规范，提供统一的身份认证、授权管理和身份治理服务。

    ## 系统特性
    - 统一身份认证与单点登录
    - 多因子认证支持
    - RBAC/ABAC混合权限模型
    - 组织架构管理
    - 应用集成与联邦身份
    - 身份治理与合规审计

    ## 认证方式
    - Bearer Token (JWT)
    - OAuth2.0
    - OpenID Connect

    ## 错误响应格式
    ```json
    {
      "code": "ERROR_CODE",
      "message": "错误描述信息",
      "timestamp": "2024-01-01T12:00:00Z",
      "path": "/api/resource"
    }
    ```
  version: 1.0.0
  contact:
    name: Qoobot开发团队
    email: dev@qoobot.com
    url: https://github.com/qoobot-com/openidaas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.idaas.example.com
    description: 生产环境
  - url: https://test-api.idaas.example.com
    description: 测试环境
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/users/{id}/roles:
    post:
      tags: [users]
      summary: 分配用户角色
      description: 为用户分配角色和权限
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '200':
          description: 分配成功
        '400':
          description: 请求参数错误
        '404':
          description: 用户或角色不存在
        '409':
          description: 角色已分配

  # 组织架构接口
  /api/organizations/departments:
    get:
      tags: [organizations]
      summary: 获取部门树
      description: 获取完整的组织架构树形结构
      security:
        - bearerAuth: []
      parameters:
        - name: parentId
          in: query
          description: 父部门ID，为空时返回根部门
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentTree'

    post:
      tags: [organizations]
      summary: 创建部门
      description: 创建新的部门
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '201':
          description: 部门创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          description: 请求参数错误
        '409':
          description: 部门编码已存在

    put:
      tags: [organizations]
      summary: 更新部门
      description: 更新部门信息
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          description: 请求参数错误
        '404':
          description: 部门不存在

    delete:
      tags: [organizations]
      summary: 删除部门
      description: 删除指定部门（需确保无下属部门和用户）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          description: 部门ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '400':
          description: 部门存在关联数据，无法删除
        '404':
          description: 部门不存在
  /api/organizations/positions:
    get:
      tags: [organizations]
      summary: 获取职位列表
      description: 查询职位信息列表
      security:
        - bearerAuth: []
      parameters:
        - name: deptId
          in: query
          description: 部门ID筛选
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'

    post:
      tags: [organizations]
      summary: 创建职位
      description: 创建新的职位
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePositionRequest'
      responses:
        '201':
          description: 职位创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '400':
          description: 请求参数错误
        '409':
          description: 职位编码已存在

    put:
      tags: [organizations]
      summary: 更新职位
      description: 更新职位信息
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePositionRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '400':
          description: 请求参数错误
        '404':
          description: 职位不存在
  /api/organizations/positions/{id}:
    get:
      tags: [organizations]
      summary: 获取职位详情
      description: 根据职位ID获取职位详细信息
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 职位ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '404':
          description: 职位不存在

    delete:
      tags: [organizations]
      summary: 删除职位
      description: 删除指定职位（需确保无用户使用）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 职位ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '400':
          description: 职位存在关联数据，无法删除
        '404':
          description: 职位不存在
...
  /api/authorization/check:
    post:
      tags: [authorization]
      summary: 权限检查
      description: 检查用户是否具有指定权限
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCheckRequest'
      responses:
        '200':
          description: 权限检查结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionCheckResponse'

  # 审计日志接口
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AssignRoleRequest:
      type: object
      required:
        - roleId
      properties:
        roleId:
          type: integer
          format: int64
          description: 角色ID
          minimum: 1
        scopeType:
          type: integer
          description: 作用域类型（1-全局，2-部门，3-项目）
          default: 1
          enum: [1, 2, 3]
        scopeId:
          type: integer
          format: int64
          description: 作用域ID
        expireTime:
          type: string
          format: date-time
          description: 过期时间
        grantReason:
          type: string
          maxLength: 500
          description: 授权原因

    # 部门管理相关
    Department:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 部门ID
        deptCode:
          type: string
          maxLength: 64
          description: 部门编码
        deptName:
          type: string
          maxLength: 128
          description: 部门名称
        parentId:
          type: integer
          format: int64
          description: 父部门ID
        levelPath:
          type: string
          maxLength: 512
          description: 层级路径
        levelDepth:
          type: integer
          description: 层级深度
        sortOrder:
          type: integer
          description: 排序序号
        managerId:
          type: integer
          format: int64
          description: 部门经理ID
        description:
          type: string
          maxLength: 1000
          description: 部门描述
        status:
          type: integer
          description: 状态（1-启用，2-禁用）
          enum: [1, 2]
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
    CreateDepartmentRequest:
      type: object
      required:
        - deptCode
        - deptName
      properties:
        deptCode:
          type: string
          minLength: 2
          maxLength: 64
          pattern: '^[A-Z0-9_]+$'
          description: 部门编码（大写字母、数字、下划线）
        deptName:
          type: string
          minLength: 2
          maxLength: 128
          description: 部门名称
        parentId:
          type: integer
          format: int64
          description: 父部门ID，0表示根部门
          default: 0
        sortOrder:
          type: integer
          description: 排序序号
          default: 0
        managerId:
          type: integer
          format: int64
          description: 部门经理ID
        description:
          type: string
          maxLength: 1000
          description: 部门描述
    UpdateDepartmentRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          description: 部门ID
          minimum: 1
        deptName:
          type: string
          minLength: 2
          maxLength: 128
          description: 部门名称
        parentId:
          type: integer
          format: int64
          description: 父部门ID
        sortOrder:
          type: integer
          description: 排序序号
        managerId:
          type: integer
          format: int64
          description: 部门经理ID
        description:
          type: string
          maxLength: 1000
          description: 部门描述
        status:
          type: integer
          description: 状态（1-启用，2-禁用）
          enum: [1, 2]

    # 职位管理相关
    CreatePositionRequest:
      type: object
      required:
        - positionCode
        - positionName
        - deptId
      properties:
        positionCode:
          type: string
          minLength: 2
          maxLength: 64
          pattern: '^[A-Z0-9_]+$'
          description: 职位编码（大写字母、数字、下划线）
        positionName:
          type: string
          minLength: 2
          maxLength: 128
          description: 职位名称
        deptId:
          type: integer
          format: int64
          description: 所属部门ID
          minimum: 1
        level:
          type: integer
          description: 职级等级
          minimum: 1
          maximum: 20
        jobGrade:
          type: string
          maxLength: 32
          description: 职级
        reportsTo:
          type: integer
          format: int64
          description: 汇报对象职位ID
        isManager:
          type: boolean
          description: 是否管理岗位
          default: false
        description:
          type: string
          maxLength: 1000
          description: 职位描述
    UpdatePositionRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          description: 职位ID
          minimum: 1
        positionCode:
          type: string
          minLength: 2
          maxLength: 64
          pattern: '^[A-Z0-9_]+$'
          description: 职位编码
        positionName:
          type: string
          minLength: 2
          maxLength: 128
          description: 职位名称
        deptId:
          type: integer
          format: int64
          description: 所属部门ID
        level:
          type: integer
          description: 职级等级
          minimum: 1
          maximum: 20
        jobGrade:
          type: string
          maxLength: 32
          description: 职级
        reportsTo:
          type: integer
          format: int64
          description: 汇报对象职位ID
        isManager:
          type: integer
          description: 是否管理岗位（0-否，1-是）
          enum: [0, 1]
        description:
          type: string
          maxLength: 1000
          description: 职位描述

    # 密码重置相关
...
    DepartmentTree:
      type: object
      properties:
        id:
          type: integer
          format: int64
        deptCode:
          type: string
        deptName:
          type: string
        parentId:
          type: integer
          format: int64
        levelPath:
          type: string
        sortOrder:
          type: integer
        children:
          type: array
          items:
            $ref: '#/components/schemas/DepartmentTree'
    Position:
      type: object
      properties:
        id:
          type: integer
          format: int64
        positionCode:
          type: string
        positionName:
          type: string
        deptId:
          type: integer
          format: int64
        level:
          type: integer
        jobGrade:
          type: string
        reportsTo:
          type: integer
          format: int64

    # 角色权限相关
...
    PermissionCheckRequest:
      type: object
      required:
        - permissions
      properties:
        userId:
          type: integer
          format: int64
          description: 用户ID，为空时检查当前用户
        permissions:
          type: array
          items:
            type: string
          description: 权限编码列表
    PermissionCheckResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        results:
          type: object
          description: 权限检查结果映射
          additionalProperties:
            type: boolean

    # 审计日志相关
...